requirements.txt
Flask==3.1.2
Flask-Limiter==3.12
Flask-Bcrypt==1.0.1
Flask-Talisman==1.1.0
Flask-Cors==5.0.0
Pillow==10.4.0
Werkzeug==3.1.0

celery==5.5.3
redis==5.0.8

transformers==4.57
torch==2.8.0
torchvision==0.23.0
sentencepiece==0.2.0
protobuf==5.27.3
accelerate==0.34.2
safetensors==0.4.4

numpy==1.26.4
pydicom==3.0.1

tqdm==4.66.5
